+++ changelog/10725.txt
@@ -1,0 +1,3 @@
+```release-note: improvement
+core (enterprise): "vault status" command works when a namespace is set.
+```
